+++ src/errors.rs
@@ -4,33 +4,29 @@ use colored::Colorize;
 use std::process::exit;
 use ErrorCode::*;
 
-#[allow(dead_code)]
 #[derive(Debug, Clone)]
 pub enum ErrorCode {
-    /// `lexer error (E0x101): unknown character: '{0}'`
-    /// - {0}: char, character
+    /// `syntax error (E0x101): malformed or unterminated char`
     E0x101,
-    /// `lexer error (E0x102): malformed or unterminated char`
+    /// `syntax error (E0x102): unterminated string`
     E0x102,
-    /// `lexer error (E0x103): unterminated string`
-    E0x103,
-    /// `lexer error (E0x104): failed to parse {1} base number '{2}'`
-    /// - {0}: string, base
-    /// - {1}: string, substring
-    E0x104,
-    /// `parser error (E0x201): unexpected token '{0}'`
+    /// `syntax error (E0x103): unexpected token '{0}'`
     /// - {0}: token
-    E0x201,
-    /// `parser error (E0x202): failed to unwrap a number '{0}'`
+    E0x103,
+    /// `syntax error (E0x104): failed to unwrap a number '{0}'`
     /// - {0}: line
     /// - {1}: number
-    E0x202,
-    /// `parser error (E0x203): failed to parse {0}`
+    E0x104,
+    /// `syntax error (E0x105): failed to parse {0}`
     /// - {0}: subject
-    E0x203,
-    /// `parser error (E0x204): expected a token '{0}'`
+    E0x105,
+    /// `syntax error (E0x106): expected a token '{0}'`
     /// - {0}: token
-    E0x204,
+    E0x106,
+    /// `sybtax error (E0x107): invalid assignment target`
+    E0x107,
+    /// `sybtax error (E0x108): expected an uppercase identifier`
+    E0x108,
     /// `runtime error (E0x301): type mismatch: expected '{0}', got '{1}'`
     /// - {0}: expected type
     /// - {1}: actual type
@@ -39,40 +35,24 @@ pub enum ErrorCode {
     E0x302,
     /// `runtime error (E0x303): return statement not within a function`
     E0x303,
-    /// `runtime error (E0x304): await statement not within an async function`
+    /// `runtime error (E0x304): failed to resolve {0}`
+    /// - {0}: subject
     E0x304,
-    /// `runtime error (E0x305): invalid function return type`
+    /// `runtime error (E0x305): '{0}' is already declared`
+    /// - {0}: subject
     E0x305,
-    /// `runtime error (E0x306): failed to resolve {0}`
-    /// - {0}: subject
+    /// `runtime error (E0x306): stack underflow`
     E0x306,
-    /// `runtime error (E0x307): '{0}' is already declared`
-    /// - {0}: subject
-    E0x307,
-    /// `runtime error (E0x308): stack underflow`
-    E0x308,
-    /// `runtime error (E0x309): stack overflow`
-    E0x309,
     /// `runtime error (E0x401): function must have one name`
     E0x401,
     /// `runtime error (E0x402): public variable must have a value`
     E0x402,
-    /// `runtime error (E0x403): invalid function body`
-    E0x403,
     /// `runtime error (E0x404): failed to create a function`
     E0x404,
     /// `runtime error (E0x405): invalid number of arguments`
     E0x405,
     /// `runtime error (E0x406): missing return statement`
     E0x406,
-    /// `runtime error (E0x407): invalid function call`
-    E0x407,
-    /// `runtime error (E0x408): failed to parse {0} literal`
-    /// - {0}: subject
-    E0x408,
-    /// `runtime error (E0x409): {0} isn't literal`
-    /// - {0}: subject
-    E0x409,
     /// `runtime error (E0x410): can not assign to an immutable variable`
     E0x410,
     /// `runtime error (E0x411): can not assign to a public variable`
@@ -114,50 +94,50 @@ impl Error {
             E0x101 => self.error(
                 101,
                 "syntax",
-                format!("unknown character '{}'", args[0]),
-                line,
-                pos,
-            ),
-            E0x102 => self.error(
-                102,
-                "syntax",
                 "malformed or unterminated char".to_string(),
                 line,
                 pos,
             ),
-            E0x103 => self.error(103, "syntax", "unterminated string".to_string(), line, pos),
-            E0x104 => self.error(
+            E0x102 => self.error(103, "syntax", "unterminated string".to_string(), line, pos),
+            E0x103 => self.error(
                 104,
-                "syntax",
-                format!("failed to parse {} base number '{}'", args[0], args[1]),
-                line,
-                pos,
-            ),
-            E0x201 => self.error(
-                201,
                 "syntax",
                 format!("unexpected token '{}'", args[0]),
                 line,
                 pos,
             ),
-            E0x202 => self.error(
-                202,
+            E0x104 => self.error(
+                105,
                 "syntax",
                 format!("failed to unwrap a number '{}'", args[0]),
                 line,
                 pos,
             ),
-            E0x203 => self.error(
-                203,
+            E0x105 => self.error(
+                106,
                 "syntax",
                 format!("failed to parse '{}'", args[0]),
                 line,
                 pos,
             ),
-            E0x204 => self.error(
-                204,
+            E0x106 => self.error(
+                108,
                 "syntax",
                 format!("expected a token '{}'", args[0]),
+                line,
+                pos,
+            ),
+            E0x107 => self.error(
+                109,
+                "syntax",
+                format!("invalid assignment target"),
+                line,
+                pos,
+            ),
+            E0x108 => self.error(
+                110,
+                "syntax",
+                format!("expected an uppercase identifier"),
                 line,
                 pos,
             ),
@@ -183,35 +163,20 @@ impl Error {
                 pos,
             ),
             E0x304 => self.error(
-                304,
-                "runtime",
-                "await statement not within an async function".to_string(),
-                line,
-                pos,
-            ),
-            E0x305 => self.error(
-                305,
-                "runtime",
-                "invalid function return type".to_string(),
-                line,
-                pos,
-            ),
-            E0x306 => self.error(
                 306,
                 "runtime",
                 format!("failed to resolve '{}'", args[0]),
                 line,
                 pos,
             ),
-            E0x307 => self.error(
+            E0x305 => self.error(
                 307,
                 "runtime",
                 format!("'{}' is already declared", args[0]),
                 line,
                 pos,
             ),
-            E0x308 => self.error(308, "runtime", "stack underflow".to_string(), line, pos),
-            E0x309 => self.error(309, "runtime", "stack overflow".to_string(), line, pos),
+            E0x306 => self.error(308, "runtime", "stack underflow".to_string(), line, pos),
             E0x401 => self.error(
                 401,
                 "runtime",
@@ -226,13 +191,7 @@ impl Error {
                 line,
                 pos,
             ),
-            E0x403 => self.error(
-                403,
-                "runtime",
-                "invalid function body".to_string(),
-                line,
-                pos,
-            ),
+
             E0x404 => self.error(
                 404,
                 "runtime",
@@ -251,27 +210,6 @@ impl Error {
                 406,
                 "runtime",
                 "missing return statement".to_string(),
-                line,
-                pos,
-            ),
-            E0x407 => self.error(
-                407,
-                "runtime",
-                "invalid function call".to_string(),
-                line,
-                pos,
-            ),
-            E0x408 => self.error(
-                408,
-                "runtime",
-                format!("failed to parse {} literal", args[0]),
-                line,
-                pos,
-            ),
-            E0x409 => self.error(
-                409,
-                "runtime",
-                format!("{} isn't literal", args[0]),
                 line,
                 pos,
             ),
@@ -402,18 +340,16 @@ impl Error {
         exit(0);
     }
 
-    #[allow(dead_code)]
-    pub fn eprintln(&self, kind: &str, code: usize, msg: String) {
-        let err_code = format!("E0x{}", code).yellow();
-        let head = format!("{} error ({}):", kind, err_code);
-        eprintln!("{} {}", head.red().bold(), msg.red());
-    }
+    // pub fn eprintln(&self, kind: &str, code: usize, msg: String) {
+    //     let err_code = format!("E0x{}", code).yellow();
+    //     let head = format!("{} error ({}):", kind, err_code);
+    //     eprintln!("{} {}", head.red().bold(), msg.red());
+    // }
 
-    #[allow(dead_code)]
-    pub fn warn(&self, msg: String) {
-        let head = "warning:".bold().yellow();
-        eprintln!("{} {}", head, msg.yellow());
-    }
+    // pub fn warn(&self, msg: String) {
+    //     let head = "warning:".bold().yellow();
+    //     eprintln!("{} {}", head, msg.yellow());
+    // }
 
     fn split_line_at_char_indices(
         &self,